pyproject.toml
[project]
name = "knowledge-base-search"
version = "0.1.0"
description = "Agentic RAG knowledge base with hybrid search using Git + Qdrant + Claude Code"
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

authors = [
    {name = "Your Name", email = "your.email@example.com"}
]

keywords = ["rag", "knowledge-base", "vector-search", "qdrant", "claude-code"]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "FlagEmbedding>=1.2.0",
    "qdrant-client>=1.9.0",
    "mcp[cli]>=1.0.0",
    "python-frontmatter>=1.1.0",
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "transformers>=4.38.0,<5.0.0",
    "anthropic>=0.18.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
